added ball class and related classes (factory & states)
changed all factories to load from image rather than filename
(plan to load a single sprite sheet with all textures; more efficient than loading the file for each factory)

added comments to explain the more obtuse code

+++ Wall.hpp
@@ -10,7 +10,7 @@ private:
 	sf::Sprite sprite;
 	Wall();
 public:
-	Wall(sf::Vector2f pos, sf::Texture & texture, int length, bool horizontal);
+	Wall(sf::Vector2f pos, sf::Texture & texture, float length, bool horizontal);
 	void update() { }
 	void draw(sf::RenderWindow & window) { window.draw(sprite); }
 	void handleEvent(sf::Event event) { }
